requirements.txt
accelerate==0.27.2
adapters==0.1.2
altair==5.3.0
altair_tiles==0.3.0
datasets==2.18.0
evaluate==0.4.1
fire==0.6.0
gensim==4.3.2
jsonlines==4.0.0
kenlm==0.2.0
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
peft==0.10.0
regex==2023.12.25
Requests==2.31.0
scikit_learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
torchsummary==1.5.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.36.0
wandb==0.16.3
